(original rev 2141) only mem-map mip file once; tweaks; 64-bit fix for stringhelper
git-svn-id: file:///tmp/current/omni/branches/staging@1901 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/utility/stringHelpers.cpp
@@ -56,7 +56,7 @@ QString StringHelpers::getStringFromStringList( const QStringList & data_set )
 	return str;
 }
 
-QString StringHelpers::commaDeliminateNumber( const int num )
+QString StringHelpers::commaDeliminateNumber( const long long num )
 {
 	QString rawNumAsStr = QString::number( num );
